adding proof of concept SQLlite database
Noggit will now generate a SqlLite database from DBC files extracted from the client directly. This gives us a generic platform to interact with for both WOTLK and SL. We Store enough information in the database to recreate the DBC files, upon building the project for patching. Features in current commit;
-Create Sqllite database with required tables
-If no database path is found, build database
-If missing table in database, add just that table (for future proofing)